+++ packages/core/src/components/Avatar/index.ts
@@ -1,5 +1,2 @@
-import Avatar from './Avatar';
-
-export * from './Avatar.type';
-export * from './Avatar';
-export default Avatar;
+export { default as Avatar } from './Avatar';
+export type { AvatarProps } from './Avatar.type';