+++ packages/core/src/components/Checkbox/index.ts
@@ -1,4 +1,3 @@
-import Checkbox from './Checkbox';
-
+export { default as Checkbox } from './Checkbox';
 export type { CheckboxProps } from './Checkbox';
-export default Checkbox;
+