+++ packages/storybook/stories/Select.stories.tsx
@@ -1,10 +1,2 @@
 // select stories for storybook
-
-import React from 'react';
-import { ComponentStory, ComponentMeta } from '@storybook/react';
-import { Select } from '@rotational/beacon-core';
-
-export default {
-  title: 'beacon/Select',
-  component: Select,
-} as ComponentMeta<typeof Select>;
+export {};