+++ packages/storybook/CHANGELOG.md
@@ -4,6 +4,17 @@
 
 ### Patch Changes
 
+- Updated dependencies [0deab6b]
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies [6b9fdf0]
+  - @rotational/beacon-foundation@3.0.0
+  - @rotational/beacon-core@3.0.0
+
+## null
+
+### Patch Changes
+
 - Updated dependencies [1d15bb3]
 - Updated dependencies [1d15bb3]
 - Updated dependencies [1d15bb3]